API CHANGE:  remove service from BaseAuth.__init__ args
It's currently impossible to construct an auth handler outside of the
context of a service's initializer (e.g. when creating a service from
inside another).  This change makes BaseAuth.service default to None,
but BaseService.__init__ will set it if it either gets auth as a keyword
arg or if it creates one of its own volition.  Be aware of this if you
set a service's auth attribute directly.