API CHANGE:  replace default_routes() with DEFAULT_ROUTES
Since Arg objects are owned by the class and not by the instance of the
class, default_routes()'s returning an instance attribute of any kind
makes args go to the wrong instance of the command when the second
instance of one is called.  We replace it with a DEFAULT_ROUTES class
attribute that can only reference instance attributes by using
callables.  It works just like it used to before ea64fcf3, but now it's
plural and usable by services and auth handlers as well.