No ParseError in python 2.6 xml module
In python 2.6, the exception class ElementTree.ParseError does not
exist. Instead the ExpatError exception is thrown. Also, the parser
might throw a SyntaxError exception. While the ParseError exception
inherits from SyntaxError, the ExpatError exception does not. In order
to fix the issue we import ExpatError only if there is no ParseError
class available and name is ParseError. When catching exceptions we
catch both ParseError and SyntaxError. Since ParseError could be an
ExpatError and may not inherit from SyntaxError, we need to explicitly
catch Syntax Error exceptions.