API CHANGE:  Refactor config, endpoint management
The Config object previously contained some global state, namely which
region and user to consider the "current" ones.  BaseService used that
region value directly to find its region name, a procedure which breaks
down when more than one region is in play or when the extended
"NAME::URL" command line syntax is used to supply a region name instead
of a configuration file.

This commit separates the state of the configuration's user and region
settings from that of services' endpoint and region name settings and
formalizes the new relationship between services and auth handlers by
handlers by replacing all of the remaining persistent references between
the two with method parameters.

The Config class is now a cloneable ConfigView, whose region and user
attributes may be set as before, but independently of other ConfigViews.
All ConfigViews share the same data.  A service's endpoint and
region_name are attributes of the BaseService class.  To allow auth
handlers to continue to obtain endpoint and region name info, __call__
was replaced with apply_to_request, which supplies a service object as
an additional argument.  Take note that the current region name will not
necessarily be the same as the name of the config view's current region.

Requesting and parsing the --region option and then applying it to the
state of the configuration forced every service class as well as every
service-less command class to supply an Arg for it and call
set_userregion one or more times to apply its state to the
configuration.  This has now been replaced with a
RegionConfigurableMixin, which add a --region Arg and an
update_config_view method that applies the value of that arg (and
optionally that of REGION_ENVVAR) to the active ConfigView.  BaseService
uses this mixin automatically; it is supplied as a separate mixin so
service-less command classes can use it as well.