Add cross-cast constructors for major base classes
Creating a request from inside an existing one requires a lot of
boilerplate parameter passing.  This commit adds "from_other" class
methods that take an instance of something and create an instance of the
new class using the same data.  This example illustrates the difference:

class SpamRequest(BaseRequest):
    def old_style(self):
        return EggsRequest(side='baked beans', config=self.config,
                           service=self.service, auth=self.auth)

    def new_style(self):
        return EggsRequest.from_other(self, side='baked beans')

Note that this will cause slicing -- it is up to you to ensure that data
from derived classes are passed correctly when they need to be.